# docs/openapi/swagger.yaml
definitions:
  controller.Login.Req:
    properties:
      Name:
        type: string
      Password:
        type: string
    required:
    - Name
    - Password
    type: object
  controller.Login.Resp:
    properties:
      token:
        type: string
    type: object
  controller.Register.Req:
    properties:
      Name:
        maxLength: 32
        minLength: 6
        type: string
      Password:
        maxLength: 32
        minLength: 8
        type: string
    required:
    - Name
    - Password
    type: object
  dm.ErrResp:
    properties:
      error:
        type: string
      id:
        type: string
    type: object
info:
  contact: {}
paths:
  /healthy:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dm.ErrResp'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dm.ErrResp'
      summary: Check the services are healthy
  /login:
    post:
      parameters:
      - description: Login User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.Login.Req'
      responses:
        "200":
          description: token
          schema:
            $ref: '#/definitions/controller.Login.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dm.ErrResp'
        "500":
          description: Internal Server Error
          schema:
            type: string
  /ping:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Check the server started
  /register:
    post:
      parameters:
      - description: Register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.Register.Req'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dm.ErrResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dm.ErrResp'
swagger: "2.0"